# unix,win: fix busy loop with zero timeout timers