@mizvekov 2023-12-16
> Can you check how this affects the million_timers benchmark?

For that benchmark, there is essentially no difference.

Follows sample output for one run, and average of 10 runs computed with hyperfine:
`hyperfine --warmup 1 'uv_run_benchmarks_a million_timers million_timers`

Old:
```
10.21 seconds total
0.44 seconds init
9.56 seconds dispatch
0.21 seconds cleanup

Time (mean ± σ):     10.299 s ±  0.010 s    [User: 1.175 s, System: 0.053 s]
Range (min … max):   10.282 s … 10.315 s    10 runs
```
New:
```
10.26 seconds total
0.47 seconds init
9.54 seconds dispatch
0.25 seconds cleanup

Time (mean ± σ):     10.299 s ±  0.009 s    [User: 1.081 s, System: 0.042 s]
Range (min … max):   10.285 s … 10.312 s    10 runs
```

Tests run on Windows 11 system with `Core i9-12900K` CPU.